Se llego a realizar compras de productos

+++ src/main/java/net/resultadofinal/micomercial/mappers/DetallecompraMapper.java
@@ -13,12 +13,12 @@ public class DetallecompraMapper implements RowMapper<DetalleCompra> {
 		DetalleCompra detallecompra = new DetalleCompra();
 		detallecompra.setCod_com(rs.getLong("cod_com"));
 		detallecompra.setCod_detcom(rs.getInt("cod_detcom"));
-		detallecompra.setCod_pro(rs.getInt("cod_pro"));
-		detallecompra.setPre_detcom(rs.getFloat("pre_detcom"));
+		detallecompra.setCod_pro(rs.getLong("cod_pro"));
+		detallecompra.setPre_detcom(rs.getBigDecimal("pre_detcom"));
 		detallecompra.setCan_detcom(rs.getInt("can_detcom"));
-		detallecompra.setDes_detcom(rs.getFloat("des_detcom"));
-		detallecompra.setSubtotDetcom(rs.getFloat("subtot_detcom"));
-		detallecompra.setTotDetcom(rs.getFloat("tot_detcom"));
+		detallecompra.setDes_detcom(rs.getBigDecimal("des_detcom"));
+		detallecompra.setSubtotDetcom(rs.getBigDecimal("subtot_detcom"));
+		detallecompra.setTotDetcom(rs.getBigDecimal("tot_detcom"));
 		return detallecompra;
 	}
 }
